explorer/puller/requirements.txt
aiohttp==3.13.3
psycopg2-binary==2.9.6
symbol-lightapi~=0.0.8
symbol-sdk-python==3.0.7
zenlog==1.1
